Fix: Server side JWT regeneration
Because middleware in Next.js only runs on the edge runtime, Prisma
will not work. Updated to use the new neon database serveless and neon
prisma adapter so that prisma calls in middleware updates the client's
JWT on the edge runtime. Now, the onboarding page and server action
will run correctly, so the client always has an up-to-date JWT.
Also added debug config, and various changes to forms, schemas, etc.