Feat: Echo, echo button spin anim, styles
Will need to make an Echo prisma schema in order to make Echos appear
at the time they were echoed, rather than appearing at the same time in
the feed as they were created. Refactor userId into username prop so
that we can indicate when someone is echoing a yap.