Updated prisma schema to include new Echo model
Needed a Echo model so that we can display Echos in the feed at the time
 they were echoed, rather than the time the original Yap was posted.
I chose to use username as the reference to the User model so that we
can spare a JOIN operation. We're always only going to need the
username when displaying an echo, not the rest of the User model.