Echo system refactored
Echos will now be shown at the time they were echoed. On the backend,
the echoYap server action is no longer dependent on state sent from the
client. To prevent instances where a user is capable of echoing a yap
more than once, the server now checks for a yap with the supplied id
and where at least one of its echoes have the username of the logged in
user. If that yap exists, the server will unecho the yap. If the does
not exist, the server will create an echo. This matches the intent of
the user as indicated by the visual state of the echo button on the
client.