Feat: Pagination for getLatestYaps
I wanted to implement infinitescroll, but I am encountering some
problems with rendering. As it exists, Posts can only be rendered on
the server. Importing them into client components throws a 'cannot
find 'fs' module error', which happens when server code gets imported
into client components. I believe this happens because of the
getLiked and getEchoed functions in the post components, which
runs a query to see if the user has already liked or echoed the post
that is being rendered. This must be done this way because Prisma ORM
does not natively support computed fields. If I wanted to implement
infinitescroll, I would have to remove these two functions, and rewrite
all the Prisma queries to use straight SQL and queryRaw, so that
computed fields for whether the user liked/echoed a post are passed as
props to the EchoYapPost or YapPost component. Thus, these two
components could be rendered on the client or server. This is too much
work and would defeat the purpose of using Prisma to begin with.
Therefore, I will use pagination with a simple button to navigate to
older posts using the next/navigate useRouter() hook and searchParams.